# depends/DNS-BUILDER/doc/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DNSBuilder API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_name}/build": {
      "post": {
        "tags": [
          "Builds"
        ],
        "summary": "Start Build",
        "operationId": "start_build_api_projects__project_name__build_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildStartRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/status": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get Build Status",
        "operationId": "get_build_status_api_builds__build_id__status_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/logs": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get Build Logs",
        "operationId": "get_build_logs_api_builds__build_id__logs_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildLogsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/artifacts": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get Build Artifacts",
        "operationId": "get_build_artifacts_api_builds__build_id__artifacts_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/builds/{build_id}/artifacts/content": {
      "get": {
        "tags": [
          "Builds"
        ],
        "summary": "Get Build Artifact Content",
        "description": "Get the file content of a specified build artifact\n\nArgs:\n    build_id: Build ID\n    path: File path relative to the output directory\n    \nReturns:\n    Plain text content of the file",
        "operationId": "get_build_artifact_content_api_builds__build_id__artifacts_content_get",
        "parameters": [
          {
            "name": "build_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Build Id"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "recursor/Dockerfile",
              "title": "Path"
            },
            "description": "recursor/Dockerfile"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactsResponse": {
        "properties": {
          "files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "ArtifactsResponse"
      },
      "BuildLogsResponse": {
        "properties": {
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs"
          },
          "last_index": {
            "type": "integer",
            "title": "Last Index"
          }
        },
        "type": "object",
        "required": [
          "logs",
          "last_index"
        ],
        "title": "BuildLogsResponse"
      },
...
      "BuildStartRequest": {
        "properties": {
          "debug": {
            "type": "boolean",
            "title": "Debug",
            "default": false
          },
          "generate_graph": {
            "type": "boolean",
            "title": "Generate Graph",
            "default": false
          }
        },
        "type": "object",
        "title": "BuildStartRequest"
      },
      "BuildStartResponse": {
        "properties": {
          "build_id": {
            "type": "string",
            "title": "Build Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "started"
          }
        },
        "type": "object",
        "required": [
          "build_id"
        ],
        "title": "BuildStartResponse"
      },
      "BuildStatusResponse": {
        "properties": {
          "project_name": {
            "type": "string",
            "title": "Project Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "start_time": {
            "type": "number",
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "project_name",
          "status",
          "start_time"
        ],
        "title": "BuildStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}